+++ src/gpodder/qml/images.py
@@ -17,6 +17,8 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 
+import urllib
+
 from PySide.QtCore import Qt
 from PySide.QtGui import QImage
 from PySide.QtDeclarative import QDeclarativeImageProvider
@@ -27,8 +29,6 @@ from gpodder import coverart
 import logging
 logger = logging.getLogger(__name__)
 
-import urllib
-
 
 class LocalCachedImageProvider(QDeclarativeImageProvider):
     IMAGE_TYPE = QDeclarativeImageProvider.ImageType.Image
@@ -38,13 +38,13 @@ class LocalCachedImageProvider(QDeclarativeImageProvider):
         self.downloader = coverart.CoverDownloader()
         self._cache = {}
 
-    def requestImage(self, id, size, requestedSize):
-        key = (id, requestedSize)
+    def requestImage(self, imageId, size, requestedSize):
+        key = (imageId, requestedSize)
         if key in self._cache:
             return self._cache[key]
 
         cover_file, cover_url, podcast_url, podcast_title = (urllib.unquote(x)
-                for x in id.split('|'))
+                for x in imageId.split('|'))
 
         def get_filename():
             return self.downloader.get_cover(cover_file, cover_url,
@@ -65,4 +65,3 @@ class LocalCachedImageProvider(QDeclarativeImageProvider):
                     Qt.SmoothTransformation)
 
         return self._cache[key]
-
